Refactor: Move BiomeDetailScreenViewModel to viewmodel package
The `BiomeDetailScreenViewModel` has been moved from `presentation/detail/biome` to a new `presentation/detail/biome/viewmodel` package.

The import statement in `BiomeDetailScreen.kt` has been updated to reflect this change.
